Add -send-nodes and -omit-implied options
With "-send-nodes false", there is no socket communication whatsoever.
This allows running the search without any visualizer listening.

The "-omit-implied" option is for replaying search from a log.  When
the search encounters a choice in the log where one child is obviously
the right choice (because of the variable's domain), we can discard
the choice point and follow the correct child.  This is done with
"-omit-implied true".  When the option is false, the branches will be
made explicitly, even though one will fail trivially (and the other one
will cause no propagation.)